requirements.txt
mysql-connector-python==8.0.31
numpy==1.24.0
pandas==1.5.2
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
